threat_intel_service/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
stix2==3.0.1
taxii2-client==2.3.0
pytz==2023.3
requests==2.31.0
pyyaml==6.0.1
marshmallow==3.20.1
stix2-patterns==2.0.0
django-filter==23.3
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
python-dateutil==2.8.2
cryptography==41.0.5
psycopg2-binary==2.9.9
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0